+++ api/src/blueprints/users.py
@@ -31,7 +31,7 @@ def create_user():
 # Read All Users (GET)
 @users_bp.route('/users', methods=['GET'])
 def get_users():
-    users = User.query.all()
+    users = db.session.scalars(db.select(User)).all()
     users_list = []
     for user in users:
         users_list.append({
@@ -50,7 +50,10 @@ def get_users():
 # Read User by ID (GET)
 @users_bp.route('/users/<int:user_id>', methods=['GET'])
 def get_user(user_id):
-    user = User.query.get_or_404(user_id)
+    user = db.session.get(User, user_id)
+    if user is None:
+        return jsonify({'error': 'User not found'}), 404
+
     user_data = {
         'user_id': user.user_id,
         'username': user.username,
@@ -67,9 +70,11 @@ def get_user(user_id):
 # Update User (PUT)
 @users_bp.route('/users/<int:user_id>', methods=['PUT'])
 def update_user(user_id):
-    user = User.query.get_or_404(user_id)
-    data = request.get_json()
+    user = db.session.get(User, user_id)
+    if user is None:
+        return jsonify({'error': 'User not found'}), 404
 
+    data = request.get_json()
     # Update user fields
     user.username = data.get('username', user.username)
     user.email = data.get('email', user.email)
@@ -89,7 +94,9 @@ def update_user(user_id):
 # Delete User (DELETE)
 @users_bp.route('/users/<int:user_id>', methods=['DELETE'])
 def delete_user(user_id):
-    user = User.query.get_or_404(user_id)
+    user = db.session.get(User, user_id)
+    if user is None:
+        return jsonify({'error': 'User not found'}), 404
     
     try:
         db.session.delete(user)
